Split MCP client connection boundaries
