+++ pyproject.toml
@@ -38,7 +38,7 @@ dependencies = [
     "httpx-sse>=0.4.0",
     "pydantic[email]>=2.6.0",
     "anyio>=4.2.0",
-    "mcp[cli]>=1.9.1",
+    "mcp[cli]>=1.27.0",
     "litellm>=1.75.2",
     "click>=8.1.7",
     "docstring_parser>=0.16",
